[VBLOCKS-4563] remove 3rd party md5 module (#334)
* feat: vendorize md5 module

* feat: remove md5 generation

* fix: rename uuid to sid

* fix: remove .only from unit test
