+++ lib/twilio/device.ts
@@ -36,6 +36,7 @@ import {
 } from './regions';
 import * as rtc from './rtc';
 import getUserMedia from './rtc/getusermedia';
+import { generateVoiceEventSid } from './sid';
 import Sound from './sound';
 import {
   isLegacyEdge,
@@ -43,7 +44,6 @@ import {
   promisifyEvents,
   queryToJson,
 } from './util';
-import { generateVoiceEventSid } from './uuid';
 
 // Placeholders until we convert the respective files to TypeScript.
 /**
