+++ tests/index.ts
@@ -110,6 +110,6 @@ require('./unit/statsMonitor');
 require('./unit/error');
 require('./unit/log');
 require('./unit/regions');
-require('./unit/uuid');
+require('./unit/sid');
 require('./unit/cdn/viewerrequest/test');
 require('./unit/cdn/viewerresponse/test');
